PRINT/INPUT finally defined.
I pulled head from ass and decided at last on a convention.

All that remains now in my initial definition is to decide whether or
not I wish to keep the Scheme convention of the appending ? for
predicates.

I'm leaning instead towards a Capital case convention, to match True and
False.

Also, I need to export quote as well, and check out eval.