Heresy is now fully functional.
It was not properly defining and providing the literals in certain macro
forms. I've also resolved some name clash issues, and switched PRINT LIT
to use write instead of print (which was naturally clashing with the
newly defined print anyway). This is a more 'CL' behavior, but at least
it pairs well with INPUT STX standing in for read.