The Y-combinator
Because I can.

Consider theory.rkt the place for stuff like combinators, a (curry ...)
procedure, (compose ...) etc. Computer science-y things and other
advanced concepts.

In the future, I may move this to a library, but for now I like the idea
of such things being a core part of the language. They're a bit hidden
away in Racket, and I think that's a shame.