Implemented basic (using ...) syntax
There is now basic file-I/O available through the (using ...) syntax.
This is essentially a parameterized block that redirects the standard
input or output (given a mode flag) to a chosen file instead. Note that
future revisions will probably extended this to include more ports, but
for now this should be enough for basic text mangling.