Reverted rnd and randomize to Racket builtins.
The xorshift* algo wasn't actually working, and my math wasn't up to
figuring out why. I've instead replaced it with what I hope to at least
be a properly parameterized approach to the Racket builtins. It's not as
'pure' as the old method, but it appears to be the more workable method
until a better algorithm can be found.