fnlet - macroizing the Y-combinator
With just a simple bit of syntax sugaring to abstract out the nested
lambdas, we now have fnlet, the named anonymous function.

Unlike a let *proc* block, which can define a recursive internal name
but is otherwise confined to its place, or can store a
non-self-referencing lambda in one of it's named arguments (which is no
different really than a local define), fnlet's can be dropped in
anywhere an fn can go.

For instance, we can do this, to map the Fibonacci sequence without
defining a named function to do it:

```
> (map (fnlet fib (n)
(select
((zero? n) 0)
((one? n) 1)
(else (+ (fib (- n 2)) (fib (- n 1))))))
(range 1 to 20))
'(1 1 2 3 5 8 13 21 34 55 89 144 233 377 610 987 1597 2584 4181 6765)
```
Bear in mind the usual caveats that of course Y is not as efficient as
traditional tail recursion (though on Racket, only just), but it may
occasionally be useful, and these little diversions are what Rule #5 is
all about.