Improve logic for `is-a?` and `thing=?`
Things now generate an internal unique ident when first created with `make-thing`, ie, when invoking a new thing by calling `describe` or `thing`. This ident is passed on to new copies of the original thing, but *not* preserved when extending/inheriting a thing (as this invokes a new call to `make-thing`.

This means that `is-a?` can now be better trusted to affirm that two Things are of the same kind, so long as they are exactly the same kind. Inheritence is still not tracked, so the following code will be false:
```racket
(describe Maybe)
(describe None extends Maybe)
(is-a? Maybe None)
```

`thing=?` behavior has now also been improved by first checking with `is-a?` that the two things are the same kind, before checking the field hashes. This means that two different objects with the same field names should not be `thing=?` any longer.