Holes do not work with monad-do
There's some kind of weird race condition shit going on with trying to
chain operations on holes. Signals wind up in the wrong places and
we're lost in async hell.

Will have to revisit the implementation at some point, to find a safer way,
but for now, hole-do is a no-go.